refactor: migrate rewards and subscriptions routes
